+++ src/modules/disk/disk.c
@@ -34,8 +34,9 @@ static void printDisk(FFDiskOptions* options, const FFDisk* disk)
     }
     else
     {
-        ffParseFormatString(&key, &options->moduleArgs.key, 1, (FFformatarg[]){
-            {FF_FORMAT_ARG_TYPE_STRBUF, &disk->mountpoint}
+        ffParseFormatString(&key, &options->moduleArgs.key, 2, (FFformatarg[]){
+            {FF_FORMAT_ARG_TYPE_STRBUF, &disk->mountpoint},
+            {FF_FORMAT_ARG_TYPE_STRBUF, &disk->name},
         });
     }